requirements.txt
# =============================================
# PRODUCTION DEPENDENCIES
# Dự án RAG Luật Việt Nam
# =============================================

# --- Core ---
numpy
pandas
tqdm
python-dotenv

# --- LangChain Ecosystem ---
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-google-genai
langchain-groq
langchain-ollama
langchain-huggingface

# --- Vector Database & Retrieval ---
chromadb
rank_bm25
sentence-transformers

# --- Document Loaders ---
unstructured
python-docx
pypdf

# --- Web Backend ---
fastapi
uvicorn[standard]

# --- Text Processing (Tiếng Việt) ---
underthesea